hosts/adk_client/pyproject.toml
[project]
name = "adk-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.14",
    "google-adk>=1.7.0",
    "google-genai>=1.26.0",
    "httpx>=0.28.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
]
